♻️ Refactor login route into FSD structure
Restructures the login page following Feature-Sliced Design (FSD) architecture for improved maintainability and testability.

Changes:
- Split monolithic login route (151 lines) into layered FSD structure
- Created LoginPage component in pages layer for composition
- Extracted business logic to useLoginView hook in widgets/login-view/model
- Created pure presentation LoginView component in widgets/login-view/ui
- Maintained 100% feature parity with existing implementation
- All functionality preserved: authentication, redirects, error handling, password visibility toggle

Architecture improvements:
- Clear separation of concerns (presentation, logic, composition)
- Testable hook-based business logic
- Reusable LoginView widget
- Consistent with existing FSD patterns (home, add-videos, player)

Verified:
- TypeScript type checking passes
- All 431 unit tests pass
- Production build succeeds
- E2E tests confirm correct behavior (login, logout, error handling)
