+++ app/pages/login/ui/LoginPage.tsx
@@ -1,0 +1,10 @@
+import { useLoginView } from '~/widgets/login-view/model/useLoginView';
+import { LoginView } from '~/widgets/login-view/ui/LoginView';
+
+export function LoginPage() {
+  const view = useLoginView();
+
+  return (
+    <LoginView {...view} />
+  );
+}
