+++ app/routes/login.tsx
@@ -1,151 +1,11 @@
-import { Eye, EyeOff, LogIn } from 'lucide-react';
-import { useEffect, useState } from 'react';
-import { useNavigate, useSearchParams } from 'react-router';
-import { Alert, AlertDescription } from '~/components/ui/alert';
-import { Button } from '~/components/ui/button';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '~/components/ui/card';
-import { Input } from '~/components/ui/input';
-import { Label } from '~/components/ui/label';
-import { useAuthStore, useIsAuthenticated } from '~/stores/auth-store';
+import type { MetaFunction } from 'react-router';
+import { LoginPage } from '~/pages/login/ui/LoginPage';
 
-export function meta() {
-  return [
-    { title: 'Login - Local Streamer' },
-    { name: 'description', content: 'Sign in to your Local Streamer account' },
-  ];
-}
+export const meta: MetaFunction = () => ([
+  { title: 'Login - Local Streamer' },
+  { name: 'description', content: 'Sign in to your Local Streamer account' },
+]);
 
-export default function Login() {
-  const navigate = useNavigate();
-  const [searchParams] = useSearchParams();
-  const isAuthenticated = useIsAuthenticated();
-  const login = useAuthStore(state => state.login);
-
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [showPassword, setShowPassword] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  // If already authenticated, redirect to intended page or home
-  useEffect(() => {
-    if (isAuthenticated) {
-      const redirectTo = searchParams.get('redirectTo') || '/';
-      navigate(redirectTo, { replace: true });
-    }
-  }, [isAuthenticated]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError(null);
-
-    // Basic validation
-    if (!email || !password) {
-      setError('Please enter both email and password');
-      return;
-    }
-
-    setLoading(true);
-
-    try {
-      const result = await login(email, password);
-
-      if (result.success) {
-        // Login successful, redirect to intended page or home
-        const redirectTo = searchParams.get('redirectTo') || '/';
-        navigate(redirectTo, { replace: true });
-      }
-      else {
-        setError(result.error || 'Login failed');
-      }
-    }
-    catch (error) {
-      console.error('Login error:', error);
-      setError('Network error occurred');
-    }
-    finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-background flex items-center justify-center p-4">
-      <Card className="w-full max-w-md">
-        <CardHeader className="text-center">
-          <div className="flex justify-center mb-4">
-            <div className="p-3 bg-primary/10 rounded-full">
-              <LogIn className="h-8 w-8 text-primary" />
-            </div>
-          </div>
-          <CardTitle className="text-2xl">Welcome Back</CardTitle>
-          <CardDescription>
-            Sign in to access your Local Streamer library
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <form onSubmit={handleSubmit} className="space-y-4">
-            {error && (
-              <Alert variant="destructive">
-                <AlertDescription>
-                  {error}
-                </AlertDescription>
-              </Alert>
-            )}
-
-            <div className="space-y-2">
-              <Label htmlFor="email">Email Address</Label>
-              <Input
-                id="email"
-                type="email"
-                value={email}
-                onChange={e => setEmail(e.target.value)}
-                placeholder="Enter your email"
-                disabled={loading}
-                autoComplete="email"
-                required
-              />
-            </div>
-
-            <div className="space-y-2">
-              <Label htmlFor="password">Password</Label>
-              <div className="relative">
-                <Input
-                  id="password"
-                  type={showPassword ? 'text' : 'password'}
-                  value={password}
-                  onChange={e => setPassword(e.target.value)}
-                  placeholder="Enter your password"
-                  disabled={loading}
-                  autoComplete="current-password"
-                  required
-                />
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="sm"
-                  className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
-                  onClick={() => setShowPassword(!showPassword)}
-                  disabled={loading}
-                >
-                  {showPassword ? (
-                    <EyeOff className="h-4 w-4" />
-                  ) : (
-                    <Eye className="h-4 w-4" />
-                  )}
-                </Button>
-              </div>
-            </div>
-
-            <Button type="submit" className="w-full" disabled={loading}>
-              {loading ? 'Signing In...' : 'Sign In'}
-            </Button>
-          </form>
-
-          <div className="mt-6 text-sm text-muted-foreground text-center">
-            <p>{`Don't have an account? Contact your administrator.`}</p>
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  );
+export default function LoginRoute() {
+  return <LoginPage />;
 }
