+++ app/widgets/login-view/ui/LoginView.tsx
@@ -1,0 +1,112 @@
+import type { FormEvent } from 'react';
+import { Eye, EyeOff, LogIn } from 'lucide-react';
+import { Alert, AlertDescription } from '~/components/ui/alert';
+import { Button } from '~/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '~/components/ui/card';
+import { Input } from '~/components/ui/input';
+import { Label } from '~/components/ui/label';
+
+export interface LoginViewProps {
+  email: string;
+  password: string;
+  showPassword: boolean;
+  loading: boolean;
+  error: string | null;
+  onEmailChange: (value: string) => void;
+  onPasswordChange: (value: string) => void;
+  onTogglePasswordVisibility: () => void;
+  onSubmit: (event: FormEvent<HTMLFormElement>) => void;
+}
+
+export function LoginView({
+  email,
+  password,
+  showPassword,
+  loading,
+  error,
+  onEmailChange,
+  onPasswordChange,
+  onTogglePasswordVisibility,
+  onSubmit,
+}: LoginViewProps) {
+  return (
+    <div className="min-h-screen bg-background flex items-center justify-center p-4">
+      <Card className="w-full max-w-md">
+        <CardHeader className="text-center">
+          <div className="flex justify-center mb-4">
+            <div className="p-3 bg-primary/10 rounded-full">
+              <LogIn className="h-8 w-8 text-primary" />
+            </div>
+          </div>
+          <CardTitle className="text-2xl">Welcome Back</CardTitle>
+          <CardDescription>
+            Sign in to access your Local Streamer library
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <form onSubmit={onSubmit} className="space-y-4">
+            {error && (
+              <Alert variant="destructive">
+                <AlertDescription>
+                  {error}
+                </AlertDescription>
+              </Alert>
+            )}
+
+            <div className="space-y-2">
+              <Label htmlFor="email">Email Address</Label>
+              <Input
+                id="email"
+                type="email"
+                value={email}
+                onChange={event => onEmailChange(event.target.value)}
+                placeholder="Enter your email"
+                disabled={loading}
+                autoComplete="email"
+                required
+              />
+            </div>
+
+            <div className="space-y-2">
+              <Label htmlFor="password">Password</Label>
+              <div className="relative">
+                <Input
+                  id="password"
+                  type={showPassword ? 'text' : 'password'}
+                  value={password}
+                  onChange={event => onPasswordChange(event.target.value)}
+                  placeholder="Enter your password"
+                  disabled={loading}
+                  autoComplete="current-password"
+                  required
+                />
+                <Button
+                  type="button"
+                  variant="ghost"
+                  size="sm"
+                  className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
+                  onClick={onTogglePasswordVisibility}
+                  disabled={loading}
+                >
+                  {showPassword ? (
+                    <EyeOff className="h-4 w-4" />
+                  ) : (
+                    <Eye className="h-4 w-4" />
+                  )}
+                </Button>
+              </div>
+            </div>
+
+            <Button type="submit" className="w-full" disabled={loading}>
+              {loading ? 'Signing In...' : 'Sign In'}
+            </Button>
+          </form>
+
+          <div className="mt-6 text-sm text-muted-foreground text-center">
+            <p>{`Don't have an account? Contact your administrator.`}</p>
+          </div>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
